Full optimization works now.
The problem was the fact that I was using the local point already
translated when calculating the jacobian for the rotation.
Also the optimizer was slow because the whole jacobian for the rotation
was scaled down by a factor of 2 for some weird reason. Needs more
investigating on the reason why but for now it's just set as a magic
number.